Add files via upload
"This repository is a collection of my frontend learning projects. Here, I’ve explored the basics of HTML and CSS through small tasks and mini-projects. Each file represents a step in my journey of understanding web concepts — from structuring pages to styling them. These are not just practice files, but building blocks that are helping me grow as a frontend developer."